Use socklen_t in socket function prototypes.
This is a POSIX requirement.

Note that this commit also adds ssize_t as a return value in some functions
(also according to POSIX) as well as size_t instead of int for input buffer
(non-socket) length parameters.

Closes #63.

+++ src/emx/src/libsocket/recv.c
@@ -36,7 +36,7 @@
 #include <InnoTekLIBC/logstrict.h>
 #include "socket.h"
 
-int recv(int socket, void *buf, int len, int flags)
+ssize_t recv(int socket, void *buf, size_t len, int flags)
 {
     LIBCLOG_ENTER("socket=%d buf=%p len=%d flags=%#x\n", socket, buf, len, flags);
     PLIBCSOCKETFH   pFHSocket = __libc_TcpipFH(socket);
