+++ src/emx/src/libsocket/send.c
@@ -36,7 +36,7 @@
 #include <InnoTekLIBC/logstrict.h>
 #include "socket.h"
 
-int send(int socket, const void *buf, int len, int flags)
+ssize_t send(int socket, const void *buf, size_t len, int flags)
 {
     LIBCLOG_ENTER("socket=%d buf=%p len=%d flags=%#x\n", socket, buf, len, flags);
     PLIBCSOCKETFH   pFHSocket = __libc_TcpipFH(socket);
