+++ src/emx/src/libsocket/sendmsg.c
@@ -36,7 +36,7 @@
 #include <InnoTekLIBC/logstrict.h>
 #include "socket.h"
 
-int sendmsg(int socket, const struct msghdr *msg, int flags)
+ssize_t sendmsg(int socket, const struct msghdr *msg, int flags)
 {
     LIBCLOG_ENTER("socket=%d msg=%p flags=%#x\n", socket, msg, flags);
     PLIBCSOCKETFH   pFHSocket = __libc_TcpipFH(socket);
